Conflicts:
	gameplay-encoder/gameplay-encoder.vcxproj.filters
	gameplay-encoder/src/AnimationChannel.cpp
	gameplay-encoder/src/AnimationChannel.h
	gameplay-encoder/src/DAEChannelTarget.cpp
	gameplay-encoder/src/DAESceneEncoder.cpp
	gameplay-encoder/src/EncoderArguments.cpp
	gameplay-encoder/src/FileIO.cpp
	gameplay-encoder/src/FileIO.h
	gameplay-encoder/src/GPBFile.cpp
	gameplay-encoder/src/Mesh.cpp
	gameplay-encoder/src/MeshSkin.cpp
	gameplay-encoder/src/MeshSkin.h
	gameplay-encoder/src/Node.cpp
	gameplay-encoder/src/Node.h
	gameplay-encoder/src/Scene.h
	gameplay-encoder/src/StringUtil.cpp
	gameplay-encoder/src/StringUtil.h
	gameplay-encoder/src/Vector2.cpp
	gameplay-encoder/src/Vector2.h
	gameplay-encoder/src/Vector3.h
	gameplay-encoder/src/Vector4.h
	gameplay-encoder/src/Vertex.cpp
	gameplay/src/Joint.cpp

+++ gameplay-encoder/src/Vector3.cpp
@@ -94,7 +94,7 @@ float Vector3::angle(const Vector3& v1, const Vector3& v2)
     float dy = v1.z * v2.x - v1.x * v2.z;
     float dz = v1.x * v2.y - v1.y * v2.x;
 
-    return atan2f(sqrtf(dx * dx + dy * dy + dz * dz) + MATH_FLOAT_SMALL, dot(v1, v2));
+    return atan2f(sqrt(dx * dx + dy * dy + dz * dz) + MATH_FLOAT_SMALL, dot(v1, v2));
 }
 
 void Vector3::add(const Vector3& v)
@@ -191,7 +191,7 @@ float Vector3::distance(const Vector3& v) const
     float dy = v.y - y;
     float dz = v.z - z;
 
-    return sqrtf(dx * dx + dy * dy + dz * dz);
+    return sqrt(dx * dx + dy * dy + dz * dz);
 }
 
 float Vector3::distanceSquared(const Vector3& v) const
@@ -215,7 +215,7 @@ float Vector3::dot(const Vector3& v1, const Vector3& v2)
 
 float Vector3::length() const
 {
-    return sqrtf(x * x + y * y + z * z);
+    return sqrt(x * x + y * y + z * z);
 }
 
 float Vector3::lengthSquared() const
